Add contract call on tokens transferred to contract
To be able to accept tokens, a contract is expected to expose a function
called `token_received`, taking a `TransferInfo` as an argument. If this
function fails, then the token contract also fails, and as such the
token will never be transferred. This allows the contract to effectively
reject the tokens it doesn't understand.